Fix Dropwizard Type code smell

+++ iotdb-core/metrics/dropwizard-metrics/src/main/java/org/apache/iotdb/metrics/dropwizard/type/DropwizardHistogramSnapshot.java
@@ -48,7 +48,7 @@ public class DropwizardHistogramSnapshot implements HistogramSnapshot {
 
   @Override
   public double getMax() {
-    return (double) snapshot.getMax();
+    return snapshot.getMax();
   }
 
   @Override
